feat(ci): split the e2e tier with a third cargo feature `e2e-live` (#502)
Issue #502's decided strategy — three lanes, split by credential need.
Adds the `e2e-live` cargo feature so the credential-requiring subset of
the L2 suite can be selected separately from the deterministic one:

- `cargo test-fast`      — fast tier (unchanged)
- `cargo test-e2e`       — lane 1: 47 files, no agent credential needed
- `cargo test-e2e-live`  — lane 1 + lane 2: all 71 files (a SUPERSET run,
                           not a live-only run)

The 24 credentialed files now open with
`#![cfg(all(feature = "e2e", feature = "e2e-live"))]`, preserving the
`unix` term on the three that carried one. No `not(feature = "e2e-live")`
term anywhere, so the combined run compiles and runs all 71.

Closes the gate-coverage hole the split would otherwise open. Under
`--features e2e` alone the 24 re-gated files compile to EMPTY CRATES —
the #407 hole reopened for a third of the tier — so the Linux `build`
job's clippy step becomes `--features e2e,e2e-live`. Measured the same
way #407 was, with an E0308 injected into tests/e2e_dispatcher_mode.rs:
`--features e2e` exited 0, `--features e2e,e2e-live` exited 101 with the
E0308. `cargo fmt --check` is unaffected (rustfmt is cfg-blind — with no
features at all it still reported a badly-formatted line inside a
live-gated file). `build-windows` / `build-macos` deliberately untouched.

No second nextest profile (issue #494): the overrides in
`.config/nextest.toml` are keyed by test-function name and are
profile-independent, so they work across all three aliases unchanged.

Also splits `tests/e2e_pane_input.rs` at test level — `prompt/pane-input/021`
(a real bash/readline pane, no credential) shared no helper and no constant
with `prompt/pane-input/022`, so the live test moved byte-identically into
`tests/e2e_pane_input_live.rs` and 021 returns to lane 1. The other two
candidates from #502 are left alone: `e2e_card_layout.rs` holds only the
live test, and `e2e_mode_indication.rs`'s two tests share three assertion
helpers.

CLAUDE.md rules 2 and 5 still describe the single-feature shape; a later
track rewrites that prose.

Refs #502, #785
