+++ tests/e2e_chain_smoke_claude.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "e2e")]
+#![cfg(all(feature = "e2e", feature = "e2e-live"))]
 
 //! L2 chain-smoke test for the real Claude Code CLI path.
 //! PRD #77 catalog `chain-smoke/claude/001`.
