+++ tests/e2e_codex_spawn_paths.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "e2e")]
+#![cfg(all(feature = "e2e", feature = "e2e-live"))]
 
 //! Synthetic recorder coverage for Wrapper-strategy launch paths.
