+++ tests/e2e_codex_worker.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "e2e")]
+#![cfg(all(feature = "e2e", feature = "e2e-live"))]
 
 //! L2 real-Codex orchestration-worker proof for PRD #20.
