+++ tests/e2e_delegate_respawn_readiness.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "e2e")]
+#![cfg(all(feature = "e2e", feature = "e2e-live"))]
 
 //! PTY-attached real-agent coverage for PRD #249's `clear = true` delegate path.
 //!
