+++ tests/e2e_delegate_work_done_chain.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "e2e")]
+#![cfg(all(feature = "e2e", feature = "e2e-live"))]
 
 //! L2 real-agent chain test for #187 / PR #188.
 //!
