+++ tests/e2e_devin_hooks.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "e2e")]
+#![cfg(all(feature = "e2e", feature = "e2e-live"))]
 
 //! PTY-attached real-Devin native-hook coverage.
 //!
