+++ tests/e2e_dispatcher_mode.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "e2e")]
+#![cfg(all(feature = "e2e", feature = "e2e-live"))]
 
 //! L2 PTY-attached reel test for PRD #220 dispatcher mode.
 //!
